Restructure scripts in Blender Docker so PYTHONPATH is not required
