+++ apps/blender/resources/images/entrypoints/scripts/render_tools/scenefileeditor.py
