+++ apps/blender/resources/images/entrypoints/scripts/render_tools/templates/blendercrop.py.template
