+++ apps/blender/resources/images/entrypoints/scripts/verifier_tools/crop_generator.py
