+++ apps/blender/resources/images/entrypoints/scripts/verifier_tools/mass_center_distance.py
