+++ apps/blender/resources/images/entrypoints/scripts/verifier_tools/wavelet.py
