+++ apps/blender/resources/images/entrypoints/verifier_entrypoint.py
@@ -1,6 +1,6 @@
 import json
 
-from verificator import verify
+from scripts.verifier_tools.verificator import verify
 
 with open('params.json', 'r') as params_file:
     params = json.load(params_file)
