+++ apps/images.ini
@@ -1,6 +1,6 @@
 golemfactory/base core/resources/images/base.Dockerfile 1.4 .
 golemfactory/nvgpu core/resources/images/nvgpu.Dockerfile 1.2 . apps.core.nvgpu.is_supported
-golemfactory/blender blender/resources/images/blender.Dockerfile 1.8 blender/resources/images/
-golemfactory/blender_verifier blender/resources/images/blender_verifier.Dockerfile 1.0 blender/resources/images/
+golemfactory/blender blender/resources/images/blender.Dockerfile igor_pr2 blender/resources/images/
+golemfactory/blender_verifier blender/resources/images/blender_verifier.Dockerfile igor_pr2 blender/resources/images/
 golemfactory/blender_nvgpu blender/resources/images/blender_nvgpu.Dockerfile 1.2 . apps.core.nvgpu.is_supported
 golemfactory/dummy dummy/resources/images/Dockerfile 1.1 dummy/resources/images
