+++ tests/golem/docker/docker-blender-cycles-task.json
@@ -11,8 +11,8 @@
         {
           "py/object": "golem.docker.image.DockerImage",
           "repository": "golemfactory/blender",
-          "tag": "1.8",
-          "name": "golemfactory/blender:1.8",
+          "tag": "igor_pr2",
+          "name": "golemfactory/blender:igor_pr2",
           "id": null
         }
       ],
@@ -56,8 +56,8 @@
     {
       "py/object": "golem.docker.image.DockerImage",
       "repository": "golemfactory/blender",
-      "tag": "1.8",
-      "name": "golemfactory/blender:1.8",
+      "tag": "igor_pr2",
+      "name": "golemfactory/blender:igor_pr2",
       "id": null
     }
   ],
