+++ tests/golem/docker/docker-blender-render-task.json
@@ -11,10 +11,10 @@
       "docker_images":[
         {
           "py/object":"golem.docker.image.DockerImage",
-          "tag":"1.8",
+          "tag":"igor_pr2",
           "id":null,
           "repository":"golemfactory/blender",
-          "name":"golemfactory/blender:1.8"
+          "name":"golemfactory/blender:igor_pr2"
         }
       ],
       "caps":[],
@@ -44,10 +44,10 @@
   "docker_images":[
     {
       "py/object":"golem.docker.image.DockerImage",
-      "tag":"1.8",
+      "tag":"igor_pr2",
       "id":null,
       "repository":"golemfactory/blender",
-      "name":"golemfactory/blender:1.8"
+      "name":"golemfactory/blender:igor_pr2"
     }
   ],
   "resolution":[
