+++ tests/golem/docker/test_docker_blender_task.py
@@ -109,7 +109,7 @@ class TestDockerBlenderCyclesTask(TestDockerBlenderTaskBase):
         assert task.header.environment == 'BLENDER'
         assert task.header.estimated_memory == 0
         assert task.docker_images[0].repository == 'golemfactory/blender'
-        assert task.docker_images[0].tag == '1.8'
+        assert task.docker_images[0].tag == 'igor_pr2'
         assert task.header.max_price == 12
         assert not task.header.signature
         assert task.listeners == []
